pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "subnet-intelligence"
version = "0.1.0"
description = "Automated Bittensor Subnet Scoring System"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "pandas>=2.2",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "schedule>=1.2",
    "pydantic>=2.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["scorer*", "api*"]
